:construction: streamline workflow
start with single dev dataset

- name parameters consistently
- model_key: save and use as given (easier for connecting
- model abbrev.: RSN, CF, DAE, VAE (make consistent)
- move helper function
