+++ project/01_1_train_CF.py
@@ -18,19 +18,9 @@
 
 # %%
 import logging
-from pathlib import Path
 from pprint import pprint
-from typing import Union, List
-
 
 import plotly.express as px
-
-# from fastai.losses import MSELossFlat
-# from fastai.learner import Learner
-
-
-import fastai
-# from fastai.tabular.all import *
 
 from fastai.basics import *
 from fastai.callback.all import *
@@ -46,25 +36,15 @@ from fastai import learner
 learner.Recorder.plot_loss = plot_loss
 # import fastai.callback.hook # Learner.summary
 
-import sklearn
-from sklearn.impute import SimpleImputer
-from sklearn.preprocessing import StandardScaler
-from sklearn.preprocessing import MinMaxScaler
 
 import vaep
-from vaep.analyzers import analyzers
 import vaep.model
 import vaep.models as models
-from vaep.models import ae
-from vaep.models import collab as vaep_collab
-from vaep.io.datasets import DatasetWithTarget
-from vaep.transform import VaepPipeline
 from vaep.io import datasplits
-# from vaep.io.dataloaders import get_dls, get_test_dl
 from vaep import sampling
 
 
-import vaep.nb as config
+import vaep.nb
 from vaep.logging import setup_logger
 logger = setup_logger(logger=logging.getLogger('vaep'))
 logger.info("Experiment 03 - Analysis of latent spaces and performance comparisions")
@@ -82,7 +62,6 @@ args = dict(globals()).keys()
 # %% tags=["parameters"]
 # files and folders
 folder_experiment:str = 'runs/experiment_03/df_intensities_proteinGroups_long_2017_2018_2019_2020_N05015_M04547/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070' # Datasplit folder with data for experiment
-folder_data:str = '' # specify data directory if needed
 file_format: str = 'pkl' # change default to pickled files
 fn_rawfile_metadata: str = 'data/files_selected_metadata.csv' # Machine parsed metadata from rawfile workflow
 # training
@@ -95,7 +74,7 @@ latent_dim:int = 10 # Dimensionality of encoding dimension (latent space of mode
 # hidden_layers:str = '128_64' # A space separated string of layers, '50 20' for the encoder, reverse will be use for decoder
 force_train:bool = True # Force training when saved model could be used. Per default re-train model
 sample_idx_position: int = 0 # position of index which is sample ID
-model_key = 'collab'
+model_key = 'CF'
 save_pred_real_na:bool=False # Save all predictions for real na
 
 # %%
@@ -203,15 +182,16 @@ test_pred_fake_na.describe()
 # larger mini-batches speed up training
 # args.batch_size_collab = args.batch_size
 
-ana_collab = models.collab.CollabAnalysis(datasplits=data,
-                                          sample_column=sample_id,
-                                          item_column=index_column, # not generic
-                                          target_column='intensity',
-                                          model_kwargs=dict(n_factors=args.latent_dim,
-                                                            y_range=(int(data.train_X.min()),
-                                                                     int(data.train_X.max())+1)
-                                                            ),
-                                          batch_size=args.batch_size_collab)
+ana_collab = models.collab.CollabAnalysis(
+    datasplits=data,
+    sample_column=sample_id,
+    item_column=index_column, # not generic
+    target_column='intensity',
+    model_kwargs=dict(n_factors=args.latent_dim,
+                    y_range=(int(data.train_X.min()),
+                                int(data.train_X.max())+1)
+                    ),
+    batch_size=args.batch_size_collab)
 
 # %%
 print("Args:")
@@ -223,8 +203,8 @@ ana_collab.model = EmbeddingDotBias.from_classes(
     classes=ana_collab.dls.classes,
     **ana_collab.model_kwargs)
 
-args.n_params_collab = models.calc_net_weight_count(ana_collab.model)
-ana_collab.params['n_parameters'] = args.n_params_collab
+args.n_params = models.calc_net_weight_count(ana_collab.model)
+ana_collab.params['n_parameters'] = args.n_params
 ana_collab.learn = Learner(dls=ana_collab.dls, model=ana_collab.model, loss_func=MSELossFlat(),
                            cbs=EarlyStoppingCallback(patience=1),
                            model_dir=args.out_models)
@@ -242,7 +222,7 @@ try:
         raise FileNotFoundError
     ana_collab.learn = ana_collab.learn.load('collab_model')
     logger.info("Loaded saved model")
-    recorder_loaded = RecorderDump.load(args.out_figures, "collab")
+    recorder_loaded = RecorderDump.load(args.out_figures, 'CF')
     logger.info("Loaded dumped figure data.")
     recorder_loaded.plot_loss()
     del recorder_loaded
@@ -250,22 +230,22 @@ except FileNotFoundError:
     suggested_lr = ana_collab.learn.lr_find()
     print(f"{suggested_lr.valley = :.5f}")
     ana_collab.learn.fit_one_cycle(args.epochs_max, lr_max=suggested_lr.valley)
-    args.epoch_collab = ana_collab.learn.epoch + 1
+    args.epoch_trained = ana_collab.learn.epoch + 1
     # ana_collab.learn.fit_one_cycle(args.epochs_max, lr_max=1e-3)
     ana_collab.model_kwargs['suggested_inital_lr'] = suggested_lr.valley
     ana_collab.learn.save('collab_model')
     fig, ax = plt.subplots(figsize=(15, 8))
-    ax.set_title('Collab loss: Reconstruction loss')
+    ax.set_title('CF loss: Reconstruction loss')
     ana_collab.learn.recorder.plot_loss(skip_start=5, ax=ax)
     recorder_dump = RecorderDump(
-        recorder=ana_collab.learn.recorder, name='collab')
+        recorder=ana_collab.learn.recorder, name='CF')
     recorder_dump.save(args.out_figures)
     del recorder_dump
     vaep.savefig(fig, name='collab_training',
                  folder=args.out_figures)
     ana_collab.model_kwargs['batch_size'] = ana_collab.batch_size
     vaep.io.dump_json(ana_collab.model_kwargs, args.out_models /
-                      TEMPLATE_MODEL_PARAMS.format("collab"))
+                      TEMPLATE_MODEL_PARAMS.format('CF'))
 
 # %% [markdown] tags=[]
 # ### Predictions
@@ -276,7 +256,7 @@ except FileNotFoundError:
 # %%
 # this could be done using the validation data laoder now
 ana_collab.test_dl = ana_collab.dls.test_dl(data.val_y.reset_index())  # test_dl is here validation data
-val_pred_fake_na['collab'], _ = ana_collab.learn.get_preds(
+val_pred_fake_na['CF'], _ = ana_collab.learn.get_preds(
     dl=ana_collab.test_dl)
 val_pred_fake_na
 
@@ -286,7 +266,7 @@ val_pred_fake_na
 
 # %%
 ana_collab.test_dl = ana_collab.dls.test_dl(data.test_y.reset_index())
-test_pred_fake_na['collab'], _ = ana_collab.learn.get_preds(dl=ana_collab.test_dl)
+test_pred_fake_na['CF'], _ = ana_collab.learn.get_preds(dl=ana_collab.test_dl)
 test_pred_fake_na
 
 # %%
@@ -298,7 +278,7 @@ if args.save_pred_real_na:
     dl_real_na = ana_collab.dls.test_dl(idx_real_na.to_frame())
     pred_real_na, _ = ana_collab.learn.get_preds(dl=dl_real_na)
     pred_real_na = pd.Series(pred_real_na, idx_real_na)
-    pred_real_na.to_csv(args.out_preds / f"pred_real_na_{args.model_key.lower()}.csv")
+    pred_real_na.to_csv(args.out_preds / f"pred_real_na_{args.model_key}.csv")
     del mask, idx_real_na, pred_real_na, dl_real_na
     # use indices of test and val to drop fake_na
     # get remaining predictions
@@ -326,7 +306,7 @@ test_pred_fake_na
 # %% [markdown]
 # ## Comparisons
 #
-# > Note: The interpolated values have less predictions for comparisons than the ones based on models (Collab, DAE, VAE)  
+# > Note: The interpolated values have less predictions for comparisons than the ones based on models (CF, DAE, VAE)  
 # > The comparison is therefore not 100% fair as the interpolated samples will have more common ones (especailly the sparser the data)  
 # > Could be changed.
 
@@ -366,20 +346,7 @@ vaep.io.dump_json(d_metrics.metrics, args.out_metrics / f'metrics_{args.model_ke
 
 
 # %% tags=[]
-def get_df_from_nested_dict(nested_dict, column_levels=['data_split', 'model', 'metric_name']):
-    metrics = {}
-    for k, run_metrics in nested_dict.items():
-        metrics[k] = vaep.pandas.flatten_dict_of_dicts(run_metrics)
-
-    metrics_dict_multikey = metrics
-
-    metrics = pd.DataFrame.from_dict(metrics, orient='index')
-    metrics.columns.names = column_levels
-    metrics.index.name = 'subset'
-    return metrics
-
-
-metrics_df = get_df_from_nested_dict(d_metrics.metrics).T
+metrics_df = models.get_df_from_nested_dict(d_metrics.metrics).T
 metrics_df
 
 # %% [markdown]
@@ -453,7 +420,6 @@ test_pred_fake_na.to_csv(args.out_preds / f"pred_test_{args.model_key}.csv")
 # ## Config
 
 # %%
-args.model_type = 'collab'
 args.dump(fname=args.out_models/ f"model_config_{args.model_key}.yaml")
 args
