+++ project/01_1_train_VAE.py
@@ -14,14 +14,13 @@
 # ---
 
 # %% [markdown]
-# # Denoising Autoencoder
+# # Variational Autoencoder
 
 # %%
 import logging
 from pathlib import Path
 from pprint import pprint
 from typing import Union, List
-
 
 import plotly.express as px
 
@@ -63,50 +62,46 @@ from vaep.io import datasplits
 # from vaep.io.dataloaders import get_dls, get_test_dl
 from vaep import sampling
 
-
 import vaep.nb as config
-from vaep.logging import setup_logger
-logger = setup_logger(logger=logging.getLogger('vaep'))
+logger = vaep.logging.setup_logger(logging.getLogger('vaep'))
 logger.info("Experiment 03 - Analysis of latent spaces and performance comparisions")
 
 figures = {}  # collection of ax or figures
 
-# %% [markdown]
-# Papermill script parameters:
 
 # %%
 # catch passed parameters
 args = None
 args = dict(globals()).keys()
 
+# %% [markdown]
+# Papermill script parameters:
 
 # %% tags=["parameters"]
 # files and folders
 folder_experiment:str = 'runs/experiment_03/df_intensities_proteinGroups_long_2017_2018_2019_2020_N05015_M04547/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070' # Datasplit folder with data for experiment
-folder_data:str = '' # specify data directory if needed
-file_format: str = 'pkl' # change default to pickled files
+file_format: str = 'pkl' # file format of create splits, default pickle (pkl)
 fn_rawfile_metadata: str = 'data/files_selected_metadata.csv' # Machine parsed metadata from rawfile workflow
 # training
-epochs_max:int = 20  # Maximum number of epochs
-# early_stopping:bool = True # Wheather to use early stopping or not
+epochs_max:int = 50  # Maximum number of epochs
 batch_size:int = 64 # Batch size for training (and evaluation)
-cuda:bool=True # Use the GPU for training?
+cuda:bool = True # Whether to use a GPU for training
 # model
 latent_dim:int = 25 # Dimensionality of encoding dimension (latent space of model)
-hidden_layers:str = '512' # A underscore separated string of layers, '128_64' for the encoder, reverse will be use for decoder
+hidden_layers:Union[int,str] = '256_128' # A underscore separated string of layers, '256_128' for the encoder, reverse will be use for decoder
 force_train:bool = True # Force training when saved model could be used. Per default re-train model
 sample_idx_position: int = 0 # position of index which is sample ID
-model_key = 'DAE'
-save_pred_real_na:bool=False # Save all predictions for real na
+model_key: str = 'VAE' # model key (lower cased version will be used for file names)
+save_pred_real_na: bool = False # Save all predictions for real na
 # metadata -> defaults for metadata extracted from machine data
-meta_date_col = None
-meta_cat_col = None
+meta_date_col: str = None # date column in meta data
+meta_cat_col: str = None # category column in meta data
 
 # %%
 # # folder_experiment = "runs/experiment_03/df_intensities_peptides_long_2017_2018_2019_2020_N05011_M42725/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070"
 # folder_experiment = "runs/experiment_03/df_intensities_evidence_long_2017_2018_2019_2020_N05015_M49321/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070"
 # latent_dim = 30
-# # epochs_max = 2
+# epochs_max = 2
 # # force_train = False
 
 # %% [markdown]
@@ -167,9 +162,11 @@ else:
 # load meta data for splits
 
 # %%
-df_meta = pd.read_csv(args.fn_rawfile_metadata, index_col=0)
-df_meta.loc[data.train_X.index.levels[0]]
-
+if args.fn_rawfile_metadata:
+    df_meta = pd.read_csv(args.fn_rawfile_metadata, index_col=0)
+    display(df_meta.loc[data.train_X.index.levels[0]])
+else:
+    df_meta = None
 
 # %% [markdown]
 # ## Initialize Comparison
@@ -181,8 +178,8 @@ df_meta.loc[data.train_X.index.levels[0]]
 #     - [x] add some additional NAs based on distribution of data
 
 # %%
-freq_peptides = sampling.frequency_by_index(data.train_X, 0)
-freq_peptides.head() # training data
+freq_feat = vaep.io.datasplits.load_freq(args.data)  
+freq_feat.head() # training data
 
 # %% [markdown]
 # ### Produce some addional fake samples
@@ -209,6 +206,7 @@ data.to_wide_format()
 args.M = data.train_X.shape[-1]
 data.train_X.head()
 
+
 # %% [markdown]
 # ### Add interpolation performance
 
@@ -219,6 +217,15 @@ test_pred_fake_na['interpolated'] = interpolated
 del interpolated
 test_pred_fake_na
 
+# %%
+# Add median pred performance
+medians_train = data.train_X.median()
+medians_train.name = 'median'
+
+val_pred_fake_na = val_pred_fake_na.join(medians_train)
+test_pred_fake_na = test_pred_fake_na.join(medians_train)
+val_pred_fake_na
+
 # %% [markdown]
 # ### Fill Validation data with potentially missing features
 
@@ -228,132 +235,139 @@ data.train_X
 # %%
 data.val_y # potentially has less features
 
-# %%
+# %% tags=[]
 data.val_y = pd.DataFrame(pd.NA, index=data.train_X.index, columns=data.train_X.columns).fillna(data.val_y)
+data.val_y
 
 # %% [markdown]
-# ## Denoising Autoencoder
+# ## Variational Autoencoder
 
 # %% [markdown]
-# ### Analysis: DataLoaders, Model, transform
+# ### Transform of data
 
 # %%
-data.train_X
-
-# %%
-data.val_y = pd.DataFrame(pd.NA, index=data.train_X.index, columns=data.train_X.columns).fillna(data.val_y)
-
-# %%
-dae_default_pipeline = sklearn.pipeline.Pipeline(
+vae_default_pipeline = sklearn.pipeline.Pipeline(
     [
         ('normalize', StandardScaler()),
         ('impute', SimpleImputer(add_indicator=False))
     ])
 
-ana_dae = ae.AutoEncoderAnalysis(train_df=data.train_X,
-                                 val_df=data.val_y,
-                                 model=ae.Autoencoder,
-                                 transform=dae_default_pipeline,
-                                 decode=['normalize'],
-                                 model_kwargs=dict(n_features=data.train_X.shape[-1],
-                                                   n_neurons=args.hidden_layers,
-                                                   last_decoder_activation=None,
-                                                   dim_latent=args.latent_dim),
-                                 bs=args.batch_size)
-args.n_params_dae = ana_dae.n_params_ae
+# %% [markdown]
+# ### Analysis: DataLoaders, Model
 
+# %%
+from torch.nn import Sigmoid
+
+ana_vae = ae.AutoEncoderAnalysis(  # datasplits=data,
+    train_df=data.train_X,
+    val_df=data.val_y,
+    model= models.vae.VAE,
+    model_kwargs=dict(n_features=data.train_X.shape[-1],
+                      h_layers=args.hidden_layers,
+                          # last_encoder_activation=None,
+                      last_decoder_activation=None,
+                      dim_latent=args.latent_dim),
+    transform=vae_default_pipeline,
+    decode=['normalize'],
+    bs=args.batch_size)
+args.n_params = ana_vae.n_params_ae
 if args.cuda:
-    ana_dae.model = ana_dae.model.cuda()
-ana_dae.model
-
-# %% [markdown]
-# ### Learner
-
-# %%
-ana_dae.learn = Learner(dls=ana_dae.dls, model=ana_dae.model,
-                        loss_func=MSELossFlat(reduction='sum'),
-                        cbs=[EarlyStoppingCallback(patience=5),
-                             ae.ModelAdapter(p=0.2)]
-                        )
-
-# %%
-ana_dae.learn.show_training_loop()
-
-# %% [markdown]
-# Adding a `EarlyStoppingCallback` results in an error.  Potential fix in [PR3509](https://github.com/fastai/fastai/pull/3509) is not yet in current version. Try again later
-
-# %%
-# learn.summary()
-
-# %%
-suggested_lr = ana_dae.learn.lr_find()
-ana_dae.params['suggested_inital_lr'] = suggested_lr.valley
-suggested_lr
-
-# %%
-vaep.io.dump_json(ana_dae.params, args.out_models / TEMPLATE_MODEL_PARAMS.format(args.model_key.lower()))
+    ana_vae.model = ana_vae.model.cuda()
+ana_vae.model
 
 # %% [markdown]
 # ### Training
-#
 
 # %%
-# papermill_description=train_dae
-ana_dae.learn.fit_one_cycle(args.epochs_max, lr_max=suggested_lr.valley)
+#self.loss_func(self.pred, *self.yb)
+
+results = []
+loss_fct = partial(models.vae.loss_fct, results=results)
+
+# %%
+# papermill_description=train_vae
+ana_vae.learn = Learner(dls=ana_vae.dls,
+                        model=ana_vae.model,
+                        loss_func=loss_fct,
+                        cbs=[ae.ModelAdapterVAE(),
+                             EarlyStoppingCallback(patience=50)
+                             ])
+
+ana_vae.learn.show_training_loop()
+# learn.summary() # see comment above under DAE
+
+# %%
+suggested_lr = ana_vae.learn.lr_find()
+ana_vae.params['suggested_inital_lr'] = suggested_lr.valley
+suggested_lr
+
+# %%
+results.clear() # reset results
 
 # %% [markdown]
-# #### Loss unnormalized
-#
-# - differences in number of total measurements not changed
+# dump model config
 
 # %%
-fig = models.plot_training_losses(learner=ana_dae.learn, name=args.model_key, folder=args.out_figures)
+# needs class as argument, not instance, but serialization needs instance
+ana_vae.params['last_decoder_activation'] = Sigmoid()
 
-# %% [markdown]
-# #### Loss normalized by total number of measurements
+vaep.io.dump_json(
+    vaep.io.parse_dict(
+        ana_vae.params, types=[
+            (torch.nn.modules.module.Module, lambda m: str(m))
+        ]),
+    args.out_models / TEMPLATE_MODEL_PARAMS.format(args.model_key))
 
-# %%
-N_train_notna = data.train_X.notna().sum().sum()
-N_val_notna = data.val_y.notna().sum().sum()
-fig = models.plot_training_losses(ana_dae.learn, args.model_key,
-                                  folder=args.out_figures,
-                                  norm_factors=[N_train_notna, N_val_notna])  # non-normalized plot of total loss
+# restore original value
+ana_vae.params['last_decoder_activation'] = Sigmoid
+
+# %% tags=[]
+ana_vae.learn.fit_one_cycle(args.epochs_max, lr_max=suggested_lr.valley)
+
 # %% [markdown]
 # Save number of actually trained epochs
 
 # %%
-args.epoch_dae = ana_dae.learn.epoch + 1
-args.epoch_dae
+args.epoch_trained = ana_vae.learn.epoch + 1
+args.epoch_trained
 
-# %% [markdown]
-# Why is the validation loss better then the training loss?
-# - during training input data is masked and needs to be reconstructed
-# - when evaluating the model, all input data is provided and only the artifically masked data is used for evaluation.
+# %%
+# results are mixed (train and validation) -> better design needed
+# in general: L_rec >> L_reg (seems so)
+# # rename _results!
+# results_train = pd.DataFrame.from_records(_results[::2], columns=['L_rec', 'L_reg'])
+# results_train.index.name = 'step'
+# results_train.plot()
+
+# %% tags=[]
+N_train_notna = data.train_X.notna().sum().sum()
+N_val_notna = data.val_y.notna().sum().sum()
+fig = models.plot_training_losses(ana_vae.learn, args.model_key, folder=args.out_figures, norm_factors=[N_train_notna, N_val_notna]) # non-normalized plot of total loss
 
 # %% [markdown]
 # ### Predictions
-#
-# - data of training data set and validation dataset to create predictions is the same as training data.
-# - predictions include real NA (which are not further compared)
-#
-# - [ ] double check ModelAdapter
-#
-# create predictiona and select for validation data
+# create predictions and select validation data predictions
 
 # %%
-ana_dae.model.eval()
-pred, target = ana_dae.get_preds_from_df(df_wide=data.train_X)  # train_X
+pred, target = res = ae.get_preds_from_df(df=data.train_X, learn=ana_vae.learn,
+                                          position_pred_tuple=0,
+                                          transformer=ana_vae.transform)
 pred = pred.stack()
-val_pred_fake_na['DAE'] = pred # model_key ?
+pred
+
+# %%
+val_pred_fake_na['VAE'] = pred # 'model_key' ?
 val_pred_fake_na
 
 # %% [markdown]
-# select predictions for test dataset
+# select test data predictions
 
-# %% tags=[]
-test_pred_fake_na['DAE'] = pred # model_key?
+# %%
+test_pred_fake_na['VAE'] = pred # model_key?
 test_pred_fake_na
 
+# %% [markdown]
+# save real na predictions
 # %%
 if args.save_pred_real_na:
     # all idx missing in training data
@@ -362,41 +376,70 @@ if args.save_pred_real_na:
     # remove fake_na idx
     idx_real_na = idx_real_na.drop(val_pred_fake_na.index).drop(test_pred_fake_na.index)
     pred_real_na = pred.loc[idx_real_na]
-    pred_real_na.to_csv(args.out_preds / f"pred_real_na_{args.model_key.lower()}.csv")
+    pred_real_na.to_csv(args.out_preds / f"pred_real_na_{args.model_key}.csv")
     del mask, idx_real_na, pred_real_na, pred
-
 
 # %% [markdown]
 # ### Plots
 #
 # - validation data
-# - [ ] add test data
 
 # %%
-# could also be a method
-ana_dae.model.cpu()
-df_latent = vaep.model.get_latent_space(ana_dae.model.encoder,
-                                            dl=ana_dae.dls.valid,
-                                            dl_index=ana_dae.dls.valid.data.index)
-df_latent
+ana_vae.model = ana_vae.model.cpu()
+df_vae_latent = vaep.model.get_latent_space(ana_vae.model.get_mu_and_logvar,
+                                            dl=ana_vae.dls.valid,
+                                            dl_index=ana_vae.dls.valid.data.index)
+df_vae_latent
 
 # %%
-ana_latent = analyzers.LatentAnalysis(df_latent,
+ana_latent_vae = analyzers.LatentAnalysis(df_vae_latent,
                                         df_meta,
                                         args.model_key,
                                         folder=args.out_figures)
 if args.meta_date_col and df_meta is not None:
-    figures[f'latent_{args.model_key.lower()}_by_date'], ax = ana_latent.plot_by_date(
+    figures[f'latent_{args.model_key}_by_date'], ax = ana_latent_vae.plot_by_date(
         args.meta_date_col)
 
 # %%
+# Could be created in data as an ID from three instrument variables
 if args.meta_cat_col and df_meta is not None:
-    figures[f'latent_{args.model_key.lower()}_by_{"_".join(args.meta_cat_col.split())}'], ax = ana_latent.plot_by_category(
+    figures[f'latent_{args.model_key}_by_{"_".join(args.meta_cat_col.split())}'], ax = ana_latent_vae.plot_by_category(
         args.meta_cat_col)
+
+# %%
+feat_freq_val = val_pred_fake_na['observed'].groupby(level=-1).count()
+feat_freq_val.name = 'freq_val'
+ax = feat_freq_val.plot.box()
+
+# %%
+# # scatter plot between overall feature freq and split freq
+# freq_feat.to_frame('overall').join(feat_freq_val).plot.scatter(x='overall', y='freq_val')
+
+# %%
+feat_freq_val.value_counts().sort_index().head() # require more than one feat?
+
+# %%
+errors_val = val_pred_fake_na.drop('observed', axis=1).sub(val_pred_fake_na['observed'], axis=0)
+errors_val = errors_val.abs().groupby(level=-1).mean()
+errors_val = errors_val.join(freq_feat).sort_values(by='freq', ascending=True)
+
+
+errors_val_smoothed = errors_val.copy() #.loc[feat_freq_val > 1]
+errors_val_smoothed[errors_val.columns[:-1]] = errors_val[errors_val.columns[:-1]].rolling(window=200, min_periods=1).mean()
+ax = errors_val_smoothed.plot(x='freq', figsize=(15,10) )
+# errors_val_smoothed
+
+# %%
+errors_val = val_pred_fake_na.drop('observed', axis=1).sub(val_pred_fake_na['observed'], axis=0)
+errors_val.abs().groupby(level=-1).agg(['mean', 'count'])
+
+# %%
+errors_val
+
 # %% [markdown]
 # ## Comparisons
 #
-# > Note: The interpolated values have less predictions for comparisons than the ones based on models (Collab, DAE, VAE)  
+# > Note: The interpolated values have less predictions for comparisons than the ones based on models (CF, DAE, VAE)  
 # > The comparison is therefore not 100% fair as the interpolated samples will have more common ones (especailly the sparser the data)  
 # > Could be changed.
 
@@ -429,27 +472,15 @@ added_metrics = d_metrics.add_metrics(test_pred_fake_na, 'test_fake_na')
 added_metrics
 
 # %% [markdown]
-# Save all metrics as json
+# ### Save all metrics as json
 
 # %% tags=[]
-vaep.io.dump_json(d_metrics.metrics, args.out_metrics / f'metrics_{args.model_key.lower()}.json')
+vaep.io.dump_json(d_metrics.metrics, args.out_metrics / f'metrics_{args.model_key}.json')
+d_metrics
 
 
 # %% tags=[]
-def get_df_from_nested_dict(nested_dict, column_levels=['data_split', 'model', 'metric_name']):
-    metrics = {}
-    for k, run_metrics in nested_dict.items():
-        metrics[k] = vaep.pandas.flatten_dict_of_dicts(run_metrics)
-
-    metrics_dict_multikey = metrics
-
-    metrics = pd.DataFrame.from_dict(metrics, orient='index')
-    metrics.columns.names = column_levels
-    metrics.index.name = 'subset'
-    return metrics
-
-
-metrics_df = get_df_from_nested_dict(d_metrics.metrics).T
+metrics_df = models.get_df_from_nested_dict(d_metrics.metrics).T
 metrics_df
 
 # %% [markdown]
@@ -512,20 +543,19 @@ fig = px.scatter(plotly_view.loc[pd.IndexSlice[:, :, subset]].stack().to_frame('
                  )
 fig.show()
 
-# %% [markdown] tags=[]
-# ## Config
-
 # %% [markdown]
 # ## Save predictions
 
 # %%
-val_pred_fake_na.to_csv(args.out_preds / f"pred_val_{args.model_key.lower()}.csv")
-test_pred_fake_na.to_csv(args.out_preds / f"pred_test_{args.model_key.lower()}.csv")
+val_pred_fake_na.to_csv(args.out_preds / f"pred_val_{args.model_key}.csv")
+test_pred_fake_na.to_csv(args.out_preds / f"pred_test_{args.model_key}.csv")
 
-# %%
-args.model_type = 'DAE'
-args.dump(fname=args.out_models/ f"model_config_{args.model_key.lower()}.yaml")
-args
+# %% [markdown] tags=[]
+# ## Config
 
 # %%
 figures # switch to fnames?
+
+# %%
+args.dump(fname=args.out_models/ f"model_config_{args.model_key}.yaml")
+args
