+++ project/config/single_dev_dataset/aggpeptides/config.yaml
@@ -2,3 +2,7 @@ config_split: config/single_dev_dataset/aggPeptides/split.yaml
 config_train: config/single_dev_dataset/aggPeptides/train.yaml
 folder_experiment: runs/experiment_03/df_intensities_peptides_long_2017_2018_2019_2020_N05011_M42725/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070
 fn_rawfile_metadata: data/files_selected_metadata.csv
+models:
+ - CF
+ - DAE
+ - VAE
