+++ project/config/single_dev_dataset/proteinGroups/config.yaml
@@ -3,3 +3,7 @@ config_train: config/single_dev_dataset/proteinGroups/train.yaml
 folder_experiment: runs/experiment_03/df_intensities_proteinGroups_long_2017_2018_2019_2020_N05015_M04547/Q_Exactive_HF_X_Orbitrap_Exactive_Series_slot_#6070
 fn_rawfile_metadata: data/files_selected_metadata.csv
 cuda: True
+models:
+ - CF
+ - DAE
+ - VAE
