+++ vaep/models/__init__.py
@@ -324,3 +324,15 @@ class Metrics():
 
     def __repr__(self):
         return pprint.pformat(self.metrics, indent=2, compact=True)
+
+
+
+def get_df_from_nested_dict(nested_dict, column_levels=['data_split', 'model', 'metric_name']):
+    metrics = {}
+    for k, run_metrics in nested_dict.items():
+        metrics[k] = vaep.pandas.flatten_dict_of_dicts(run_metrics)
+
+    metrics = pd.DataFrame.from_dict(metrics, orient='index')
+    metrics.columns.names = column_levels
+    metrics.index.name = 'subset'
+    return metrics
